practice.cpp

+++ 2Linear Search/2Optimized.cpp
@@ -5,18 +5,22 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-int linearSearchOptimized(const vector<int>& arr, int target) {
-    for (int i = 0; i < (int)arr.size(); ++i) {
-        if (arr[i] == target) return i;
-        if (arr[i] > target) break; // useful when the array is sorted
+int linearSearchOptimized(const vector<int> &arr, int target)
+{
+    for (int i = 0; i < arr.size(); ++i)
+    {
+        if (arr[i] == target)
+            return i;
+        if (arr[i] > target)
+            break; // useful when the array is sorted asc
     }
     return -1;
 }
 
-int main() {
+int main()
+{
     vector<int> arr = {1, 3, 5, 7, 9};
     int target = 7;
     cout << "Index: " << linearSearchOptimized(arr, target) << "\n";
     return 0;
 }
-
